+++ modules/Grades/Configuration.php
@@ -21,12 +21,13 @@ $grades = DBGet( "SELECT cp.TITLE AS CP_TITLE,c.TITLE AS COURSE_TITLE,cp.COURSE_
 FROM REPORT_CARD_GRADES rcg,COURSE_PERIODS cp,COURSES c
 WHERE cp.COURSE_ID=c.COURSE_ID
 AND cp.TEACHER_ID='" . User( 'STAFF_ID' ) . "'
+AND cp.SCHOOL_ID='" . UserSchool() . "'
 AND cp.SCHOOL_ID=rcg.SCHOOL_ID
 AND cp.SYEAR=rcg.SYEAR
 AND cp.SYEAR='" . UserSyear() . "'
 AND rcg.GRADE_SCALE_ID=cp.GRADE_SCALE_ID
 AND cp.GRADE_SCALE_ID IS NOT NULL
-AND DOES_BREAKOFF='Y'
+AND cp.DOES_BREAKOFF='Y'
 ORDER BY rcg.BREAK_OFF IS NOT NULL DESC,rcg.BREAK_OFF DESC,rcg.SORT_ORDER DESC", array(), array( 'COURSE_PERIOD_ID' ) );
 
 echo '<form action="Modules.php?modname=' . $_REQUEST['modname'] . '" method="POST">';
@@ -37,18 +38,29 @@ echo '<br />';
 
 PopTable( 'header', _( 'Configuration' ) );
 
-echo '<fieldset>';
-//FJ add translation
-//FJ css WPadmin
-echo '<legend>' . _( 'Assignments' ) . '</legend>';
-echo '<table>';
+// Allow Edit fields for teachers.
+$_ROSARIO['allow_edit'] = true;
+
+echo '<fieldset><legend>' . _( 'Assignments' ) . '</legend><table>';
 
 if ( count( $grades ) )
 {
 	//if ( ! $gradebook_config['ROUNDING'])
 	//	$gradebook_config['ROUNDING'] = 'NORMAL';
-	//FJ add <label> on radio
-	echo '<tr><td><table><tr><td colspan="4"><b>' . _( 'Score Rounding' ) . '</b></td></tr><tr><td><label><input type="radio" name="values[ROUNDING]" value=UP' . (  ( $gradebook_config['ROUNDING'] == 'UP' ) ? ' checked' : '' ) . '>&nbsp;' . _( 'Up' ) . '</label></td><td><label><input type="radio" name="values[ROUNDING]" value=DOWN' . (  ( $gradebook_config['ROUNDING'] == 'DOWN' ) ? ' checked' : '' ) . '>&nbsp;' . _( 'Down' ) . '</label></td><td><label><input type="radio" name="values[ROUNDING]" value="NORMAL"' . (  ( $gradebook_config['ROUNDING'] == 'NORMAL' ) ? ' checked' : '' ) . '>&nbsp;' . _( 'Normal' ) . '</label></td><td><label><input type="radio" name="values[ROUNDING]" value="' . (  ( $gradebook_config['ROUNDING'] == '' ) ? ' checked' : '' ) . '">&nbsp;' . _( 'None' ) . '</label></td></tr></table></td></tr>';
+
+	$rounding_options = array(
+		'UP' => _( 'Up' ),
+		'DOWN' => _( 'Down' ),
+		'NORMAL' => _( 'Normal' ),
+	);
+
+	echo '<tr><td>' . RadioInput(
+		$gradebook_config['ROUNDING'],
+		'values[ROUNDING]',
+		_( 'Score Rounding' ),
+		$rounding_options,
+		_( 'None' )
+	) . '</td></tr>';
 }
 
 if ( empty( $gradebook_config['ASSIGNMENT_SORTING'] ) )
@@ -56,59 +68,86 @@ if ( empty( $gradebook_config['ASSIGNMENT_SORTING'] ) )
 	$gradebook_config['ASSIGNMENT_SORTING'] = 'ASSIGNMENT_ID';
 }
 
-echo '<tr><td><table><tr><td colspan="3"><b>' . _( 'Assignment Sorting' ) .
-'</b></td></tr><tr class="st"><td><label><input type="radio" name="values[ASSIGNMENT_SORTING]" value="ASSIGNMENT_ID"' .
-( $gradebook_config['ASSIGNMENT_SORTING'] == 'ASSIGNMENT_ID' ? ' checked' : '' ) . '>&nbsp;' .
-_( 'Newest First' ) .
-'</label></td><td><label><input type="radio" name="values[ASSIGNMENT_SORTING]" value="DUE_DATE"' .
-(  $gradebook_config['ASSIGNMENT_SORTING'] == 'DUE_DATE' ? ' checked' : '' ) . '>&nbsp;' .
-_( 'Due Date' ) .
-'</label></td><td><label><input type="radio" name="values[ASSIGNMENT_SORTING]" value=ASSIGNED_DATE' .
-( $gradebook_config['ASSIGNMENT_SORTING'] == 'ASSIGNED_DATE' ? ' checked' : '' ) . '>&nbsp;' .
-_( 'Assigned Date' ) . '</label></td></tr></table></td></tr>';
+$sorting_options = array(
+	'ASSIGNMENT_ID' => _( 'Newest First' ),
+	'DUE_DATE' => _( 'Due Date' ),
+	'ASSIGNED_DATE' => _( 'Assigned Date' ),
+);
 
-//FJ add <label> on checkbox
-echo '<tr><td><label><input type="checkbox" name="values[WEIGHT]" value="Y"' .
-( ! empty( $gradebook_config['WEIGHT'] ) ? ' checked' : '' ) . '> ' .
-_( 'Weight Grades' ) . '</label></td></tr>';
+echo '<tr><td>' . RadioInput(
+	$gradebook_config['ASSIGNMENT_SORTING'],
+	'values[ASSIGNMENT_SORTING]',
+	_( 'Assignment Sorting' ),
+	$sorting_options,
+	false
+) . '</td></tr>';
 
-echo '<tr><td><label><input type="checkbox" name="values[DEFAULT_ASSIGNED]" value="Y"' .
-( ! empty( $gradebook_config['DEFAULT_ASSIGNED'] ) ? ' checked' : '' ) . '> ' .
-_( 'Assigned Date defaults to today' ) . '</label></td></tr>';
+echo '<tr><td><hr />' . CheckboxInput(
+	$gradebook_config['WEIGHT'],
+	'values[WEIGHT]',
+	_( 'Weight Grades' )
+) . '</td></tr>';
 
-echo '<tr><td><label><input type="checkbox" name="values[DEFAULT_DUE]" value="Y"' .
-( ! empty( $gradebook_config['DEFAULT_DUE'] ) ? ' checked' : '' ) . '> ' . _( 'Due Date defaults to today' ) .
-'</label></td></tr>';
+echo '<tr><td>' . CheckboxInput(
+	$gradebook_config['DEFAULT_ASSIGNED'],
+	'values[DEFAULT_ASSIGNED]',
+	_( 'Assigned Date defaults to today' )
+) . '</td></tr>';
 
-echo '<tr><td><label><input type="checkbox" name="values[LETTER_GRADE_ALL]" value="Y"' .
-( ! empty( $gradebook_config['LETTER_GRADE_ALL'] ) ? ' checked' : '' ) . '> ' .
-_( 'Hide letter grades for all gradebook assignments' ) . '</label></td></tr>';
+echo '<tr><td>' . CheckboxInput(
+	$gradebook_config['DEFAULT_DUE'],
+	'values[DEFAULT_DUE]',
+	_( 'Due Date defaults to today' )
+) . '</td></tr>';
 
-echo '<tr><td><input type="text" name="values[LETTER_GRADE_MIN]" value="' .
-$gradebook_config['LETTER_GRADE_MIN'] . '" size="3" maxlength="3" /> ' .
-_( 'Minimum assignment points for letter grade' ) . '</td></tr>';
+echo '<tr><td>' . CheckboxInput(
+	$gradebook_config['LETTER_GRADE_ALL'],
+	'values[LETTER_GRADE_ALL]',
+	_( 'Hide letter grades for all gradebook assignments' )
+) . '</td></tr>';
 
-echo '<tr><td><input type="text" name="values[ANOMALOUS_MAX]" value="' .
-( ! empty( $gradebook_config['ANOMALOUS_MAX'] ) ? $gradebook_config['ANOMALOUS_MAX'] : '100' ) . '" size="3" maxlength="3" /> % ' .
-_( 'Allowed maximum percent in Anomalous grades' ) . '</td></tr>';
+echo '<tr><td><hr />' . TextInput(
+	$gradebook_config['LETTER_GRADE_MIN'],
+	'values[LETTER_GRADE_MIN]',
+	_( 'Minimum assignment points for letter grade' ),
+	'size="3" maxlength="3"'
+) . '</td></tr>';
 
-echo '<tr><td><input type="text" name="values[LATENCY]" value="' .
-round( $gradebook_config['LATENCY'] ) . '" size="3" maxlength="3" /> ' .
-_( 'Days until ungraded assignment grade appears in Parent/Student gradebook views' ) . '</td></tr>';
+$anomalous_max_value = ( ! empty( $gradebook_config['ANOMALOUS_MAX'] ) ? $gradebook_config['ANOMALOUS_MAX'] : '100' );
 
-echo '</table>';
-echo '</fieldset><br />';
+$anomalous_max_value = array(
+	$anomalous_max_value,
+	$anomalous_max_value . '%'
+);
+
+echo '<tr><td>' . TextInput(
+	$anomalous_max_value,
+	'values[ANOMALOUS_MAX]',
+	_( 'Allowed maximum percent in Anomalous grades' ),
+	'size="3" maxlength="3"'
+) . '</td></tr>';
+
+echo '<tr><td>' . TextInput(
+	(string) round( $gradebook_config['LATENCY'] ),
+	'values[LATENCY]',
+	_( 'Days until ungraded assignment grade appears in Parent/Student gradebook views' ),
+	'size="3" maxlength="3"'
+) . '</td></tr>';
+
+
+echo '</table></fieldset><br />';
 
 if ( $RosarioModules['Eligibility'] )
 {
-	echo '<fieldset>';
-	echo '<legend>' . _( 'Eligibility' ) . '</legend>';
-	echo '<table>';
-	echo '<tr><td><label><input type="checkbox" name="values[ELIGIBILITY_CUMULITIVE]" value="Y"' .
-	( ! empty( $gradebook_config['ELIGIBILITY_CUMULITIVE'] ) ? ' checked' : '' ) . '>&nbsp;' .
-	_( 'Calculate Eligibility using Cumulative Semester Grades' ) . '</label></td></tr>';
-	echo '</table>';
-	echo '</fieldset><br />';
+	echo '<fieldset><legend>' . _( 'Eligibility' ) . '</legend><table>';
+
+	echo '<tr><td>' . CheckboxInput(
+		$gradebook_config['ELIGIBILITY_CUMULITIVE'],
+		'values[ELIGIBILITY_CUMULITIVE]',
+		_( 'Calculate Eligibility using Cumulative Semester Grades' )
+	) . '</td></tr>';
+
+	echo '</table></fieldset><br />';
 }
 
 $comment_codes_RET = DBGet( "SELECT rccs.ID,rccs.TITLE,rccc.TITLE AS CODE_TITLE
@@ -119,9 +158,7 @@ ORDER BY rccc.SORT_ORDER,rccs.SORT_ORDER,rccs.ID,rccc.ID", array(), array( 'ID' 
 
 if ( $comment_codes_RET )
 {
-	echo '<fieldset>';
-	echo '<legend>' . _( 'Final Grades' ) . '</legend>';
-	echo '<table class="col1-align-right">';
+	echo '<fieldset><legend>' . _( 'Final Grades' ) . '</legend><table class="col1-align-right">';
 
 	foreach ( (array) $comment_codes_RET as $id => $comments )
 	{
@@ -130,15 +167,14 @@ if ( $comment_codes_RET )
 		foreach ( (array) $comments as $key => $val )
 		{
 			echo '<option value="' . $val['CODE_TITLE'] . '"' .
-			( $val['CODE_TITLE'] == $gradebook_config['COMMENT_' . $id] ? ' selected' : '' ) . '>' .
-			$val['CODE_TITLE'];
+				( $val['CODE_TITLE'] == $gradebook_config['COMMENT_' . $id] ? ' selected' : '' ) . '>' .
+				$val['CODE_TITLE'];
 		}
 
 		echo '</select></td><td>' . sprintf( _( 'Default %s comment code' ), $comments[1]['TITLE'] ) . '</td></tr>';
 	}
 
-	echo '</table>';
-	echo '</fieldset><br />';
+	echo '</table></fieldset><br />';
 }
 
 /*
@@ -154,9 +190,7 @@ $grades[ $course_period_id ][ $i ] = $grades[ $course_period_id ][ $i ]['TITLE']
 
 if ( count( $grades ) )
 {
-	echo '<fieldset>';
-	echo '<legend>' . _( 'Score Breakoff Points' ) . '</legend>';
-	echo '<table><tr><td>';
+	echo '<fieldset><legend>' . _( 'Score Breakoff Points' ) . '</legend><table>';
 
 	foreach ( (array) $grades as $course_period_id => $cp_grades )
 	{
@@ -185,14 +219,12 @@ if ( count( $grades ) )
 			}
 		}
 
-		$table .= '</tr>';
-		$table .= '</table>';
-		echo $table;
-		echo '</td></tr><tr><td>';
+		$table .= '</tr></table>';
+
+		echo '<tr><td>' . $table . '</td></tr>';
 	}
 
-	echo '</td></tr></table>';
-	echo '</fieldset><br />';
+	echo '</table></fieldset><br />';
 }
 
 $year = DBGet( "SELECT TITLE,MARKING_PERIOD_ID,DOES_GRADES
@@ -216,9 +248,7 @@ $quarters = DBGet( "SELECT TITLE,MARKING_PERIOD_ID,PARENT_ID,DOES_GRADES
 	AND SCHOOL_ID='" . UserSchool() . "'
 	ORDER BY SORT_ORDER", array(), array( 'PARENT_ID' ) );
 
-echo '<fieldset>';
-echo '<legend>' . _( 'Final Grading Percentages' ) . '</legend>';
-echo '<table>';
+echo '<fieldset><legend>' . _( 'Final Grading Percentages' ) . '</legend><table>';
 
 foreach ( (array) $semesters as $sem )
 {
@@ -246,6 +276,7 @@ foreach ( (array) $semesters as $sem )
 		}
 
 		$table .= '</tr></table>';
+
 		echo '<tr><td>' . $table . '</td></tr>';
 	}
 }
@@ -274,7 +305,7 @@ if ( $year[1]['DOES_GRADES'] === 'Y' )
 		{
 			$table .= '<td><span class="nobr">' . $sem['TITLE'] . '&nbsp;</span><br />';
 			$table .= '<input type="text" name="values[FY-' . $sem['MARKING_PERIOD_ID'] . ']" value="' .
-			$gradebook_config['FY-' . $sem['MARKING_PERIOD_ID']] . '" size="3" maxlength="6" />%</td>';
+				$gradebook_config['FY-' . $sem['MARKING_PERIOD_ID']] . '" size="3" maxlength="6" />%</td>';
 
 			$total += $gradebook_config['FY-' . $sem['MARKING_PERIOD_ID']];
 		}
@@ -286,11 +317,11 @@ if ( $year[1]['DOES_GRADES'] === 'Y' )
 	}
 
 	$table .= '</tr></table>';
+
 	echo '<tr><td>' . $table . '</td></tr>';
 }
 
-echo '</table>';
-echo '</fieldset>';
+echo '</table></fieldset>';
 
 PopTable( 'footer' );